Remove cascading dependency from Grouped/CoGrouped, seal the traits
